# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Private Moments API
  version: 0.1.0
  description: API contract for the Mac server used by the iOS app and admin UI.
servers:
  - url: http://127.0.0.1:3210
security:
  - bearerAuth: []
paths:
...
  /api/v1/ai/media-summary:
    post:
      summary: Generate or regenerate an AI summary for one audio/video media file.
      description: >
        The Mac server transcribes the stored audio/video media file locally, then
        calls the configured external Chat Completions provider with the internal
        transcript. Provider credentials stay server-side, and the iOS client only
        receives generated summary metadata and lightweight failure status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaSummaryRequest"
      responses:
        "200":
          description: Summary status or generated summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaSummaryResponse"
  /api/v1/ai/media-summary/{summaryId}:
    delete:
      summary: Soft delete one generated AI media summary.
      parameters:
        - name: summaryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted summary metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaSummaryResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MediaSummaryRequest:
      type: object
      required:
        - postId
        - mediaId
      properties:
        postId:
          type: string
          format: uuid
        mediaId:
          type: string
          format: uuid
        forceRegenerate:
          type: boolean
          default: false
    MediaSummaryResponse:
      type: object
      required:
        - summary
      properties:
        summary:
          $ref: "#/components/schemas/AISummary"
    AISummary:
      type: object
      required:
        - id
        - postId
        - mediaId
        - status
        - keyPoints
        - sections
        - documentBlocks
        - promptVersion
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        mediaId:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - transcribing
            - summarizing
            - ready
            - failed
            - deleted
        format:
          type:
            - string
            - "null"
          enum:
            - sentence
            - summary_with_points
            - sectioned
            - document
            - null
        language:
          type:
            - string
            - "null"
        overview:
          type:
            - string
            - "null"
        keyPoints:
          type: array
          items:
            type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/AISummarySection"
        summaryText:
          type:
            - string
            - "null"
        documentTitle:
          type:
            - string
            - "null"
          description: Optional generated display title for v2 document-style summaries.
        oneLiner:
          type:
            - string
            - "null"
          description: Short top-level summary for v2 document-style summaries.
        documentBlocks:
          type: array
          description: Native block model rendered by iOS as a Markdown-like summary document. Empty for legacy summaries.
          items:
            $ref: "#/components/schemas/AISummaryDocumentBlock"
        inputTranscriptLength:
          type:
            - integer
            - "null"
          description: Nullable when no usable transcript was produced or older generated records used a no-transcript path.
        inputDurationSeconds:
          type:
            - number
            - "null"
        promptVersion:
          type: string
        provider:
          type:
            - string
            - "null"
        model:
          type:
            - string
            - "null"
        errorCode:
          type:
            - string
            - "null"
        errorMessage:
          type:
            - string
            - "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type:
            - string
            - "null"
          format: date-time
    AISummarySection:
      type: object
      required:
        - heading
        - bullets
      properties:
        heading:
          type: string
        bullets:
          type: array
          items:
            type: string
    AISummaryDocumentBlock:
      type: object
      required:
        - kind
        - level
        - text
        - items
      properties:
        kind:
          type: string
          enum:
            - heading
            - paragraph
            - bullets
            - numbered_list
            - ai_suggested
        level:
          type: integer
          enum:
            - 0
            - 1
            - 2
          description: Heading depth for heading blocks; 0 for non-heading blocks.
        text:
          type: string
        items:
          type: array
          items:
            type: string
...